IA-integration/ia-consumer/requirements.txt
pandas
openpyxl
scikit-learn
joblib
paho-mqtt